Add "adapter" to filename of each Compose adapter file
