+++ deferred-resources-compose-adapter/src/androidTest/java/com/backbase/deferredresources/compose/DeferredBooleanAdapterTest.kt
@@ -1,9 +1,8 @@
 package com.backbase.deferredresources.compose
 
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.backbase.deferredresources.DeferredColor
+import com.backbase.deferredresources.DeferredBoolean
 import com.backbase.deferredresources.compose.test.GenericValueNode
 import com.backbase.deferredresources.compose.test.R
 import com.backbase.deferredresources.compose.test.TestTag
@@ -13,25 +12,13 @@ import org.junit.Rule
 import org.junit.Test
 
 @OptIn(ExperimentalComposeAdapter::class)
-internal class DeferredColorTest {
+internal class DeferredBooleanAdapterTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
 
-    @Test fun constructWithComposeColor_resolvesExpectedValue() {
-        val deferred = DeferredColor.Constant(Color(0xff00ff00))
-        composeTestRule.setContent {
-            GenericValueNode(
-                value = deferred.resolve(),
-                modifier = TestTagModifier,
-            )
-        }
-
-        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(Color.Green)
-    }
-
     @Test fun resolve_withLocalContext_returnsExpectedValue() {
-        val deferred = DeferredColor.Resource(R.color.blue)
+        val deferred = DeferredBoolean.Resource(R.bool.testBool)
         composeTestRule.setContent {
             GenericValueNode(
                 value = deferred.resolve(),
@@ -39,6 +26,6 @@ internal class DeferredColorTest {
             )
         }
 
-        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(Color.Blue)
+        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(true)
     }
 }
