+++ deferred-resources-compose-adapter/src/androidTest/java/com/backbase/deferredresources/compose/DeferredColorAdapterTest.kt
@@ -1,8 +1,9 @@
 package com.backbase.deferredresources.compose
 
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.backbase.deferredresources.DeferredInteger
+import com.backbase.deferredresources.DeferredColor
 import com.backbase.deferredresources.compose.test.GenericValueNode
 import com.backbase.deferredresources.compose.test.R
 import com.backbase.deferredresources.compose.test.TestTag
@@ -12,13 +13,13 @@ import org.junit.Rule
 import org.junit.Test
 
 @OptIn(ExperimentalComposeAdapter::class)
-internal class DeferredIntegerTest {
+internal class DeferredColorAdapterTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
 
-    @Test fun resolve_withLocalContext_returnsExpectedValue() {
-        val deferred = DeferredInteger.Resource(R.integer.five)
+    @Test fun constructWithComposeColor_resolvesExpectedValue() {
+        val deferred = DeferredColor.Constant(Color(0xff00ff00))
         composeTestRule.setContent {
             GenericValueNode(
                 value = deferred.resolve(),
@@ -26,6 +27,18 @@ internal class DeferredIntegerTest {
             )
         }
 
-        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(5)
+        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(Color.Green)
+    }
+
+    @Test fun resolve_withLocalContext_returnsExpectedValue() {
+        val deferred = DeferredColor.Resource(R.color.blue)
+        composeTestRule.setContent {
+            GenericValueNode(
+                value = deferred.resolve(),
+                modifier = TestTagModifier,
+            )
+        }
+
+        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(Color.Blue)
     }
 }
