+++ deferred-resources-compose-adapter/src/androidTest/java/com/backbase/deferredresources/compose/DeferredDrawableAdapterTest.kt
@@ -17,7 +17,7 @@ import org.junit.Rule
 import org.junit.Test
 
 @OptIn(ExperimentalComposeAdapter::class)
-internal class DeferredDrawableTest {
+internal class DeferredDrawableAdapterTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
