+++ deferred-resources-compose-adapter/src/androidTest/java/com/backbase/deferredresources/compose/DeferredFormattedStringAdapterTest.kt
@@ -2,7 +2,7 @@ package com.backbase.deferredresources.compose
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.backbase.deferredresources.DeferredBoolean
+import com.backbase.deferredresources.DeferredFormattedString
 import com.backbase.deferredresources.compose.test.GenericValueNode
 import com.backbase.deferredresources.compose.test.R
 import com.backbase.deferredresources.compose.test.TestTag
@@ -12,20 +12,20 @@ import org.junit.Rule
 import org.junit.Test
 
 @OptIn(ExperimentalComposeAdapter::class)
-internal class DeferredBooleanTest {
+internal class DeferredFormattedStringAdapterTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
 
     @Test fun resolve_withLocalContext_returnsExpectedValue() {
-        val deferred = DeferredBoolean.Resource(R.bool.testBool)
+        val deferred = DeferredFormattedString.Resource(R.string.formattedString)
         composeTestRule.setContent {
             GenericValueNode(
-                value = deferred.resolve(),
+                value = deferred.resolve("localized"),
                 modifier = TestTagModifier,
             )
         }
 
-        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(true)
+        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals("This is localized text.")
     }
 }
