+++ deferred-resources-compose-adapter/src/androidTest/java/com/backbase/deferredresources/compose/DeferredIntegerAdapterTest.kt
@@ -2,7 +2,7 @@ package com.backbase.deferredresources.compose
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.backbase.deferredresources.DeferredFormattedString
+import com.backbase.deferredresources.DeferredInteger
 import com.backbase.deferredresources.compose.test.GenericValueNode
 import com.backbase.deferredresources.compose.test.R
 import com.backbase.deferredresources.compose.test.TestTag
@@ -12,20 +12,20 @@ import org.junit.Rule
 import org.junit.Test
 
 @OptIn(ExperimentalComposeAdapter::class)
-internal class DeferredFormattedStringTest {
+internal class DeferredIntegerAdapterTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
 
     @Test fun resolve_withLocalContext_returnsExpectedValue() {
-        val deferred = DeferredFormattedString.Resource(R.string.formattedString)
+        val deferred = DeferredInteger.Resource(R.integer.five)
         composeTestRule.setContent {
             GenericValueNode(
-                value = deferred.resolve("localized"),
+                value = deferred.resolve(),
                 modifier = TestTagModifier,
             )
         }
 
-        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals("This is localized text.")
+        composeTestRule.onNodeWithTag(TestTag).assertGenericValueEquals(5)
     }
 }
