+++ deferred-resources-compose-adapter/src/main/java/com/backbase/deferredresources/compose/DeferredDimensionAdapter.kt
