+++ deferred-resources-compose-adapter/src/main/java/com/backbase/deferredresources/compose/DeferredTextAdapter.kt
